[#30] Setup initial (broken) structure of the gameserver.
The gameserver includes cardscheme as library allowing us to work on the
interpreter isolated.

Even after hours of teaking maven I cannot get it to compile and
resolve the dependencies as needed. For some reason IntelliJ gets it
right but I don't know how I can teach maven to do the same.