[#31] Implement error messages for name clashes of functions, loops etc.
Now the parser detects such name clashes and discards the source code.

This however would be a great example of finding an error in the parser
from which we could easily recover and still explore the rest of the
program. Unfortunately, at currently the interpreter is not set up to do
that.