[#92] Refactor code by removing null handling
In a review it was questioned whether or not a variable read can be
safely unwrapped. And while it could, one needed to read more
code and the unwrapping was a pain to look at.

So now I refactored the variable to not be optional and some other
null unwrappings.